requirements.txt

pandas
requests
azure-storage-blob
